+++ example/codeceptjs/edit_todos_test.js
@@ -10,10 +10,10 @@ Before(async (I, TodosPage) => {
 
 Scenario('Edited todo is saved on blur', async (I, TodosPage) => {
     I.say('Given I have some todos')
-  
+
     I.say('When I edit the first todo')
     await TodosPage.editNthTodo(1, 'boom')
-  
+
     I.say('Then I see that the todo text has been changed')
     await TodosPage.seeNthTodoEquals(1, 'boom')
 
@@ -26,6 +26,11 @@ Scenario('Delete todos', async (I, TodosPage) => {
     TodosPage.deleteNthTodo(1)
 
     I.say('Then the todo should disappear from the list')
-    TodosPage.seeNumberOfTodos(2)    
+    TodosPage.seeNumberOfTodos(2)
 })
-  
+
+const inputs = [1,2,3];
+
+inputs.forEach((input) => {
+  Scenario(`Have multiple ${input} data`, () => {
+})});